Split SymPy normalization tests by ownership
